pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gaggle"
dynamic = ["version"]
description = "AI-Powered Agile Development Team"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Dan Oliver", email = "dan@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    # Data models and validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    # CLI and configuration
    "typer>=0.9.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "aiohttp>=3.13.2",
    "boto3>=1.40.69",
    # GitHub and LLM integrations
    "PyGithub>=2.1.0",
    "anthropic>=0.34.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    
    # Code quality
    "ruff>=0.3.0",
    "black>=24.2.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
]

api = [
    # API framework (optional)
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "gunicorn>=21.2.0",
]

[project.urls]
Homepage = "https://github.com/danoliver1/gaggle"
Documentation = "https://gaggle.readthedocs.io"
Repository = "https://github.com/danoliver1/gaggle"
Issues = "https://github.com/danoliver1/gaggle/issues"

[project.scripts]
gaggle = "gaggle.main:app"

[tool.hatch.version]
path = "src/gaggle/__init__.py"

[tool.ruff]
target-version = "py310"
line-length = 88
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings  
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "A",    # flake8-builtins
    "T20",  # flake8-print
]
ignore = [
    "E501",  # Line too long (handled by formatter)
    "T201",  # Print found (we use print in CLI)
    "A003",  # Class attribute is shadowing a python builtin
]

fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["gaggle"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/gaggle --cov-report=html --cov-report=term-missing"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]
